refactor(msgraph-effect-client): align domains with MS Graph workloads, kebab-case with .service.ts/.schema.ts suffixes
Restructures src/ to use lowercase kebab-case folder names and .service.ts/.schema.ts file suffixes matching MS Graph API workload groupings. Updates all ServiceMap.Service identifier strings to include the workload folder segment (e.g. MsGraph/Users/UsersService) and rewrites index.ts with proper domain-grouped exports.
