+++ packages/msgraph-effect-client/src/Auth/Permissions.ts
