--- packages/msgraph-effect-client/src/MsGraph/index.ts
@@ -1,23 +1,0 @@
-export * from './Auth/MsGraphAuth';
-export * from './Auth/MsGraphAuthConfig';
-export * from './Auth/Permissions';
-export * from './Auth/TokenCache';
-export * from './Calendar/CalendarService';
-export * from './Drive/DriveService';
-export * from './Errors/errors';
-export * from './Groups/GroupsService';
-export * from './Http/BatchRequest';
-export * from './Http/MsGraphHttpClient';
-export * from './Mail/MailService';
-export * from './Planner/PlannerService';
-export * from './Schemas/Common';
-export * from './Schemas/DriveItem';
-export * from './Schemas/Event';
-export * from './Schemas/Group';
-export * from './Schemas/Message';
-export * from './Schemas/OData';
-export * from './Schemas/PlannerTask';
-export * from './Schemas/Team';
-export * from './Schemas/User';
-export * from './Teams/TeamsService';
-export * from './Users/UsersService';
