+++ packages/msgraph-effect-client/src/groups/group.schema.ts
