+++ packages/msgraph-effect-client/src/index.ts
@@ -1,1 +1,35 @@
-export * from './MsGraph/index';
+// Errors
+
+// Auth
+export * from './auth/ms-graph-auth.js';
+export * from './auth/ms-graph-auth-config.js';
+export * from './auth/permissions.js';
+export * from './auth/token-cache.js';
+// Calendar  (/users/{id}/events)
+export * from './calendar/calendar.service.js';
+export * from './calendar/event.schema.js';
+export * from './errors/errors.js';
+export * from './files/drive-item.schema.js';
+// Files  (/drives, /me/drive)
+export * from './files/files.service.js';
+export * from './groups/group.schema.js';
+// Groups  (/groups)
+export * from './groups/groups.service.js';
+export * from './http/batch-request.js';
+// Http
+export * from './http/ms-graph-http-client.js';
+// Mail  (/users/{id}/messages)
+export * from './mail/mail.service.js';
+export * from './mail/message.schema.js';
+// Planner  (/planner)
+export * from './planner/planner.service.js';
+export * from './planner/planner-task.schema.js';
+// Shared Schemas
+export * from './schemas/common.schema.js';
+export * from './schemas/odata.schema.js';
+export * from './teams/team.schema.js';
+// Teams  (/teams)
+export * from './teams/teams.service.js';
+export * from './users/user.schema.js';
+// Users  (/users, /me)
+export * from './users/users.service.js';
