+++ packages/msgraph-effect-client/src/schemas/common.schema.ts
