+++ packages/msgraph-effect-client/src/schemas/odata.schema.ts
