+++ packages/msgraph-effect-client/src/users/user.schema.ts
